move kernel_primitives to pten
