+++ paddle/fluid/operators/fused/attn_gemm.h
@@ -15,8 +15,8 @@ limitations under the License. */
 #include "paddle/fluid/platform/float16.h"
 
 #include "paddle/fluid/operators/elementwise/elementwise_op_broadcast.cu.h"
-#include "paddle/fluid/operators/kernel_primitives/kernel_primitives.h"
 #include "paddle/fluid/operators/reduce_ops/reduce_op.cu.h"
+#include "paddle/pten/kernels/primitive/kernel_primitives.h"
 
 namespace paddle {
 namespace operators {
