+++ paddle/pten/kernels/primitive/datamover_primitives.h
@@ -13,18 +13,68 @@
 // limitations under the License.
 
 #pragma once
-#include "xpu/kernel/cluster_header.h"
-#include "xpu/kernel/debug.h"
-#include "xpu/kernel/math.h"
+#ifdef PADDLE_WITH_CUDA
+#include <cuda.h>
+#include <cuda_fp16.h>
+#endif
+#ifdef PADDLE_WITH_HIP
+#include <hip/hip_fp16.h>
+#endif
+#include "paddle/pten/core/ddim.h"
 
-namespace paddle {
-namespace operators {
-namespace kernel_primitives {
+namespace pten {
+namespace kps {
 namespace details {
+
+#define INT_BITS 32
 
 template <typename T, int VecSize>
 struct alignas(sizeof(T) * VecSize) VectorType {
   T val[VecSize];
+};
+/**
+ * Fast division : Replace division in CUDA with multiplication to improve
+ * kernel performance.
+ * 1. Complete the division calculation on the CPU, and record the calculation
+ * results by using the divider and shift_val.
+ * 2. Set the divisor on the GPU through Div() to complete the calculation.
+ */
+struct FastDivMod {
+  // 1st value represents the result of input number divides by recorded divisor
+  // 2nd value represents the result of input number modulo by recorded divisor
+  using DivModT = VectorType<uint32_t, 2>;
+
+  FastDivMod() {}
+  HOSTDEVICE FastDivMod(uint32_t d) : divisor(d) {
+    static_assert(sizeof(unsigned int) == 4,
+                  "Only Support 32-bit unsigned int.");
+
+    for (shift_val = 0; shift_val < INT_BITS; ++shift_val) {
+      auto shift_limit = 1 << shift_val;
+      if (shift_limit >= divisor) break;
+    }
+    uint64_t long_one = 1;
+    uint64_t temp_div =
+        ((long_one << INT_BITS) * ((long_one << shift_val) - divisor)) /
+            divisor +
+        1;
+    multiplier = temp_div;
+  }
+
+  __device__ __forceinline__ uint32_t Div(uint32_t n) const {
+    uint32_t t = __umulhi(n, multiplier);
+    return (t + n) >> shift_val;
+  }
+
+  __device__ __forceinline__ DivModT Divmod(uint32_t n) const {
+    uint32_t q = Div(n);
+    DivModT result = {q, n - q * divisor};
+    return result;
+  }
+
+  int32_t divisor;
+  int32_t shift_val;
+  uint32_t multiplier;
 };
 
 /**
@@ -32,51 +82,40 @@ struct alignas(sizeof(T) * VecSize) VectorType {
  * index of the output data. if input or output shape is [dim0, dim1] then dims
  * must be [dim1, dim0].
  */
-#pragma pack(4)
 template <int kDims>
 struct BroadcastConfig {
-  int strides_in[framework::DDim::kMaxRank];
-  int strides_out[framework::DDim::kMaxRank];
-  int in_dim[framework::DDim::kMaxRank];
-
+  FastDivMod divmoders[kDims];
+  uint32_t strides[pten::framework::DDim::kMaxRank];
   HOSTDEVICE BroadcastConfig() {}
 
   HOSTDEVICE BroadcastConfig(const std::vector<int64_t>& out_dims,
                              const std::vector<int64_t>& in_dims,
                              int dim_size) {
-    std::vector<int> strides_in_tmp;
-    std::vector<int> strides_out_tmp;
-    std::vector<int> dim_tmp;
-    strides_in_tmp.resize(dim_size, 1);
-    strides_out_tmp.resize(dim_size, 1);
-    dim_tmp.resize(dim_size, 1);
-    for (int i = 1; i < dim_size; i++) {
-      strides_in_tmp[i] = strides_in_tmp[i - 1] * in_dims[i - 1];
-      strides_out_tmp[i] = strides_out_tmp[i - 1] * out_dims[i - 1];
+    std::vector<uint32_t> strides_in;
+    std::vector<FastDivMod> divmoders_in;
+    // for divmoders
+    divmoders_in.resize(dim_size);
+    for (int i = 0; i < dim_size; ++i) {
+      divmoders_in[i] = FastDivMod(out_dims[i]);
+    }
+    // for strides
+    strides_in.resize(dim_size, 1);
+    for (int i = 0; i < dim_size; ++i) {
+      strides_in[i] = in_dims[i] == 1 ? 0 : strides_in[i];
+      strides_in[i] = (i != 0 && strides_in[i] != 0)
+                          ? std::accumulate(in_dims.begin(),
+                                            in_dims.begin() + i,
+                                            1,
+                                            std::multiplies<int64_t>())
+                          : strides_in[i];
     }
 
-    for (int i = 0; i < dim_size; i++) {
-      dim_tmp[i] = in_dims[i];
-    }
-
-    memcpy(strides_in, strides_in_tmp.data(), kDims * sizeof(int));
-    memcpy(strides_out, strides_out_tmp.data(), kDims * sizeof(int));
-    memcpy(in_dim, dim_tmp.data(), kDims * sizeof(int));
-  }
-
-  __device__ inline int operator()(int index_output) const {
-    int index_src = 0;
-#pragma unroll
-    for (int i = kDims - 1; i >= 0; --i) {
-      int tmp_index = (index_output / strides_out[i]);
-      index_output = index_output - tmp_index * strides_out[i];
-      index_src += (tmp_index % in_dim[i]) * strides_in[i];
-    }
-    return index_src;
+    memcpy(strides, strides_in.data(), kDims * sizeof(uint32_t));
+    memcpy(divmoders, divmoders_in.data(), kDims * sizeof(FastDivMod));
   }
 };
-#pragma pack()
 
+#undef INT_BITS
 }  // namespace details
 
 /**
@@ -88,11 +127,11 @@ struct BroadcastConfig {
  * Ty: The type of data that needs to be stored in registers.
  * NX: The number of data columns loaded by each thread.
  * NY: The number of data rows loaded by each thread.
- * BlockSize: Identifies the current device thread index method. For xpu,
- * core_id() is used as the index.
+ * BlockSize: Identifies the current device thread index method. For GPU,
+ * threadIdx.x is used as the thread index. Currently only GPU was supported.
  * IsBoundary: Indicates whether to perform block access storage out-of-bounds
  * judgment. When the number of data processed by the block is less than
- * NX x NY x core_num(), boundary judgment is required to avoid memory access
+ * NX x NY x blockDim, boundary judgment is required to avoid memory access
  * crossing the boundary.
  *
  * @param：
@@ -105,24 +144,29 @@ struct BroadcastConfig {
  * stride_nx: Each read one element stride stride_nx elements in the last dim.
  * stride_ny: Each read one element stride stride_ny elements in the first dim.
  */
-template <typename Tx, typename Ty, int NX, int NY, int BlockSize,
+template <typename Tx,
+          typename Ty,
+          int NX,
+          int NY,
+          int BlockSize,
           bool IsBoundary = false>
-__device__ __inline__ void ReadData(Ty* dst, const Tx _global_ptr_* src,
-                                    int size_nx, int size_ny, int stride_nx,
-                                    int stride_ny) {
-  int thread_offset = core_id();
+__device__ __forceinline__ void ReadData(Ty* dst,
+                                         const Tx* __restrict__ src,
+                                         int size_nx,
+                                         int size_ny,
+                                         int stride_nx,
+                                         int stride_ny) {
+  int thread_offset = threadIdx.x;
   int left_size_nx = size_nx - thread_offset;
-  __local__ Tx in_temp[1];
+
   // Each branch is added for better performance
   if (NX == 1 && NY == 1) {  // for NX == 1 and NY == 1
     if (IsBoundary) {
       if (left_size_nx > 0) {
-        GM2LM(src + thread_offset, in_temp, sizeof(Tx));
-        dst[0] = static_cast<Ty>(in_temp[0]);
+        dst[0] = static_cast<Ty>(src[thread_offset]);
       }
     } else {
-      GM2LM(src + thread_offset, in_temp, sizeof(Tx));
-      dst[0] = static_cast<Ty>(in_temp[0]);
+      dst[0] = static_cast<Ty>(src[thread_offset]);
     }
   } else if (NX == 1) {  // for NX == 1 and NY != 1
 #pragma unroll
@@ -132,8 +176,7 @@ __device__ __inline__ void ReadData(Ty* dst, const Tx _global_ptr_* src,
           break;
         }
       }
-      GM2LM(src + thread_offset + idy * stride_ny, in_temp, sizeof(Tx));
-      dst[idy] = static_cast<Ty>(in_temp[0]);
+      dst[idy] = static_cast<Ty>(src[thread_offset + idy * stride_ny]);
     }
   } else if (NY == 1) {  // for NY == 1 and NX != 1
 #pragma unroll
@@ -143,22 +186,25 @@ __device__ __inline__ void ReadData(Ty* dst, const Tx _global_ptr_* src,
           break;
         }
       }
-      GM2LM(src + thread_offset + idx * stride_nx, in_temp, sizeof(Tx));
-      dst[idx] = static_cast<Ty>(in_temp[0]);
+      dst[idx] = static_cast<Ty>(src[thread_offset + idx * stride_nx]);
     }
   } else {  // for NX != 1 and NY != 1
 #pragma unroll
     for (int idx = 0; idx < NX; ++idx) {
+      if (IsBoundary) {
+        if (idx * stride_nx >= left_size_nx) {
+          break;
+        }
+      }
 #pragma unroll
       for (int idy = 0; idy < NY; ++idy) {
         if (IsBoundary) {
-          if (idy * stride_ny >= size_ny || idx * stride_nx >= left_size_nx) {
+          if (idy * stride_ny >= size_ny) {
             break;
           }
         }
-        int fix = thread_offset + idx * stride_nx + idy * stride_ny;
-        GM2LM(src + fix, in_temp, sizeof(Tx));
-        dst[idy * NX + idx] = static_cast<Ty>(in_temp[0]);
+        dst[idy * NX + idx] = static_cast<Ty>(
+            src[thread_offset + idx * stride_nx + idy * stride_ny]);
       }
     }
   }
@@ -176,7 +222,7 @@ __device__ __inline__ void ReadData(Ty* dst, const Tx _global_ptr_* src,
  * init_data: Initial value.
  */
 template <typename T, int NX>
-__device__ __inline__ void Init(T* dst, T init_data) {
+__device__ __forceinline__ void Init(T* dst, T init_data) {
 #pragma unroll
   for (int i = 0; i < NX; i++) {
     dst[i] = init_data;
@@ -192,11 +238,11 @@ __device__ __inline__ void Init(T* dst, T init_data) {
  * T: The type of data.
  * NX: Each thread load NX data from global memory continuously.
  * NY: Each thread need to load NY rows, only NY = 1 was supported.
- * BlockSize: Identifies the current device thread index method. For xpu,
- * core_id() is used as the index.
+ * BlockSize: Identifies the current device thread index method. For GPU,
+ * threadIdx.x is used as the thread index. Currently only GPU was supported.
  * IsBoundary: Whether to make an out-of-bounds judgment on access to memory.
  * When the number of data processed by this block is less than
- * NX x NY x core_num(), boundary judgment is required to avoid memory access
+ * NX x NY x blockDim.x, boundary judgment is required to avoid memory access
  * crossing the boundary.
  *
  * @param：
@@ -205,20 +251,34 @@ __device__ __inline__ void Init(T* dst, T init_data) {
  * size: The current block needs to load size data continuously.
  */
 template <typename T, int NX, int NY, int BlockSize, bool IsBoundary = false>
-__device__ __inline__ void ReadData(T* dst, const T _global_ptr_* src,
-                                    int num) {
-  int thread_offset = core_id() * NX;
-  __local__ T in_temp[1];
-  if (IsBoundary) {  // core_num() * NX > num
+__device__ __forceinline__ void ReadData(T* dst,
+                                         const T* __restrict__ src,
+                                         int num) {
+  if (IsBoundary) {  // blockDim.x * NX > num
+    int thread_offset = threadIdx.x * NX;
 #pragma unroll
     for (int idx = 0; idx < NX; ++idx) {
       if (idx + thread_offset < num) {
-        GM2LM(src + thread_offset + idx, in_temp, sizeof(T));
-        dst[idx] = in_temp[0];
+        dst[idx] = src[thread_offset + idx];
       }
     }
-  } else {  // core_num() * NX < num
-    GM2LM(src + thread_offset, dst, NX * sizeof(T));
+  } else {  // blockDim,x * NX < num
+    constexpr int kVectorSize = (NX % 4 == 0) ? 4 : (NX % 2 == 0) ? 2 : 1;
+    constexpr int kVectorsPerThread = NX / kVectorSize;
+    int thread_offset = threadIdx.x * kVectorsPerThread;
+
+    using VecType = details::VectorType<T, kVectorSize>;
+    const VecType* vec_input = reinterpret_cast<const VecType*>(src);
+    VecType vec_temp[kVectorsPerThread];
+
+#pragma unroll
+    for (int i = 0; i < kVectorsPerThread; ++i) {
+      vec_temp[i] = vec_input[thread_offset + i];
+#pragma unroll
+      for (int idx = 0; idx < NX; ++idx) {
+        dst[idx] = *(reinterpret_cast<T*>(vec_temp) + idx);
+      }
+    }
   }
 }
 
@@ -229,34 +289,40 @@ __device__ __inline__ void ReadData(T* dst, const T _global_ptr_* src,
  * T: The type of data stored in the global memory.
  * NX: The number of data columns loaded by each thread.
  * NY: The number of data rows loaded by each thread.
- * BlockSize: Identifies the current device thread index method. For xpu,
- * core_id() is used as the index.
+ * BlockSize: Identifies the current device thread index method. For GPU,
+ * threadIdx.x is used as the thread index. Currently only GPU was supported.
  * Rank: The shape size of out. eg in[1, 35], out[32, 35] then shape size is 2.
  * IsBoundary: Indicates whether to perform block access storage out-of-bounds
  * judgment. When the number of data processed by the block is less than
- * NX x NY x core_num(), boundary judgment is required to avoid memory access
+ * NX x NY x blockDim.x, boundary judgment is required to avoid memory access
  * crossing the boundary.
  *
  * @param：
  * dst: The register pointer of the thread, the size is NX * NY.
- * src: Raw input data pointer of kernel.
- * block_offset: Data offset of this block, core_num() *  cluster_id() * NX;
+ * src: The original input data pointer of this kernel.
+ * block_offset: The data offset of this block, blockDim.x * blockIdx.x * NX.
  * config: Calculation configuration of broadcast. It is used to calculate the
  * coordinate mapping relationship between output data and input data.
  * total_num_output: Total number of original output.
  * stride_nx: Each read one element stride stride_nx elements in the last dim.
  * stride_ny: Each read one element stride stride_ny elements in the first dim.
  */
-template <typename T, int NX, int NY, int BlockSize, int Rank,
+template <typename T,
+          int NX,
+          int NY,
+          int BlockSize,
+          int Rank,
           bool IsBoundary = false>
-__device__ __inline__ void ReadDataBc(T* dst, const T _global_ptr_* src,
-                                      uint32_t block_offset,
-                                      details::BroadcastConfig<Rank> config,
-                                      int total_num_output, int stride_nx,
-                                      int stride_ny) {
-  uint32_t thread_offset = block_offset + core_id();
+__device__ __forceinline__ void ReadDataBc(
+    T* dst,
+    const T* __restrict__ src,
+    uint32_t block_offset,
+    details::BroadcastConfig<Rank> config,
+    int total_num_output,
+    int stride_nx,
+    int stride_ny) {
+  uint32_t thread_offset = block_offset + threadIdx.x;
   uint32_t index_src = 0;
-  __local__ T in_temp[1];
 
 #pragma unroll
   for (int ny = 0; ny < NY; ++ny) {
@@ -265,13 +331,17 @@ __device__ __inline__ void ReadDataBc(T* dst, const T _global_ptr_* src,
       uint32_t index_output = thread_offset + ny * stride_ny + nx * stride_nx;
       index_src = 0;
       if (IsBoundary) {
-        if (index_output >= (uint32_t)total_num_output) {
+        if (index_output >= total_num_output) {
           break;
         }
       }
-      index_src = config(index_output);
-      GM2LM(src + index_src, in_temp, sizeof(T));
-      dst[nx + ny * NX] = in_temp[0];
+#pragma unroll
+      for (int i = 0; i < Rank; ++i) {
+        auto fast_divmoder = config.divmoders[i].Divmod(index_output);
+        index_output = fast_divmoder.val[0];
+        index_src += fast_divmoder.val[1] * config.strides[i];
+      }
+      dst[nx + ny * NX] = src[index_src];
     }
   }
 }
@@ -283,18 +353,18 @@ __device__ __inline__ void ReadDataBc(T* dst, const T _global_ptr_* src,
  * T: The type of data.
  * NX: The number of data columns loaded by each thread.
  * NY: The number of data rows loaded by each thread.
- * BlockSize: Identifies the current device thread index method. For xpu,
- * core_id() is used as the index.
+ * BlockSize: Identifies the current device thread index method. For GPU,
+ * threadIdx.x is used as the thread index. Currently only GPU was supported.
  * Rank: The shape size of out. eg in[1, 35], out[32, 35] then shape size is 2.
  * IsBoundary: Indicates whether to perform block access storage out-of-bounds
  * judgment. When the number of data processed by the block is less than
- * NX x NY x core_num(), boundary judgment is required to avoid memory access
+ * NX x NY x blockDim.x, boundary judgment is required to avoid memory access
  * crossing the boundary.
  *
  * @param：
  * dst: The register pointer of the thread, the size is NX * NY.
  * src: The input data pointer of this block.
- * block_offset: The data offset of this block, blockDim.x * cluster_id() * NX.
+ * block_offset: The data offset of this block, blockDim.x * blockIdx.x * NX.
  * index_cal: Calculation configuration of Reduce. It is used to calculate the
  * coordinate mapping relationship between output data and input data.
  * size_nx: The current block needs to load size_nx columns of data, this
@@ -307,23 +377,33 @@ __device__ __inline__ void ReadDataBc(T* dst, const T _global_ptr_* src,
  * reduce_last_dim: Used to indicate whether the dimension of reduce contains
  * the lowest dimension.
  */
-template <typename T, int NX, int NY, int BlockSize, int Rank,
-          typename IndexCal, bool IsBoundary = false>
-__device__ __inline__ void ReadDataReduce(T* dst, const T _global_ptr_* src,
-                                          int block_offset,
-                                          const IndexCal& index_cal,
-                                          int size_nx, int size_ny,
-                                          int stride_nx, int stride_ny,
-                                          bool reduce_last_dim) {
-  __local__ Tx in_temp[1];
+template <typename Tx,
+          typename Ty,
+          int NX,
+          int NY,
+          int BlockSize,
+          int Rank,
+          typename IndexCal,
+          typename Functor,
+          bool IsBoundary = false>
+__device__ __forceinline__ void ReadDataReduce(Ty* dst,
+                                               const Tx* __restrict__ src,
+                                               int block_offset,
+                                               const IndexCal& index_cal,
+                                               int size_nx,
+                                               int size_ny,
+                                               int stride_nx,
+                                               int stride_ny,
+                                               Functor func,
+                                               bool reduce_last_dim) {
   int thread_offset = 0;
   int left_idx = 0;
   if (reduce_last_dim) {
-    thread_offset = core_id();
-    left_idx = 0;
+    thread_offset = threadIdx.x;
+    left_idx = threadIdx.y;
   } else {
-    thread_offset = 0;
-    left_idx = 0;
+    thread_offset = threadIdx.y;
+    left_idx = threadIdx.x;
   }
 
   if (NX == 1) {
@@ -335,8 +415,7 @@ __device__ __inline__ void ReadDataReduce(T* dst, const T _global_ptr_* src,
         }
       }
       uint32_t index_src = index_cal(thread_offset + block_offset);
-      GM2LM(src + index_src, in_temp, sizeof(Tx));
-      dst[ny] = static_cast<Ty>(func(in_temp[0]));
+      dst[ny] = static_cast<Ty>(func(src[index_src]));
       thread_offset += stride_ny;
     }
   } else {
@@ -351,15 +430,15 @@ __device__ __inline__ void ReadDataReduce(T* dst, const T _global_ptr_* src,
           }
         }
         uint32_t index_src = index_cal(thread_offset + block_offset);
-        GM2LM(src + index_src, in_temp, sizeof(Tx));
-        dst[nx + ny * NX] = static_cast<Ty>(func(in_temp[0]));
+        dst[nx + ny * NX] = static_cast<Ty>(func(src[index_src]));
         thread_offset += stride_ny;
       }
     }
   }
 }
+
 /**
- * @brief Write 1D data from registers to global memory. When IsBoundary = true
+ * @brief Write 2D data from registers to global memory. When IsBoundary = true
  * and (NX % 4 == 0 or Nx % 2 == 0), the data will be vectorized to improve the
  * data loading efficiency
  *
@@ -367,11 +446,11 @@ __device__ __inline__ void ReadDataReduce(T* dst, const T _global_ptr_* src,
  * T: The type of data.
  * NX: The number of data continuously writed by each thread.
  * NY: The number of data rows loaded by each thread, only NY = 1 was supported.
- * BlockSize: Identifies the current device thread index method. For xpu,
- * core_id() is used as the index.
+ * BlockSize: Identifies the current device thread index method. For GPU,
+ * threadIdx.x is used as the thread index. Currently only GPU was supported.
  * IsBoundary: Indicates whether to perform block access storage out-of-bounds
  * judgment. When the number of data processed by the block is less than
- * NX x NY x core_num(), boundary judgment is required to avoid memory access
+ * NX x NY x blockDim.x, boundary judgment is required to avoid memory access
  * crossing the boundary.
  *
  * @param：
@@ -379,21 +458,32 @@ __device__ __inline__ void ReadDataReduce(T* dst, const T _global_ptr_* src,
  * src: The register pointer, the size is NX * NY.
  * size: The current block needs to load size elements continuously.
  */
-
-template <typename T, int NX, int NY, int BlockSize, bool IsBoundary>
-__device__ void WriteData(T _global_ptr_* dst, const T* src, int num) {
-  int thread_offset = core_id() * NX;
-  __local__ T in_temp[1];
-  if (IsBoundary) {  // core_num() * NX > num
+template <typename T, int NX, int NY, int BlockSize, bool IsBoundary = false>
+__device__ __forceinline__ void WriteData(T* dst,
+                                          T* __restrict__ src,
+                                          int num) {
+  if (IsBoundary) {
+    int thread_offset = threadIdx.x * NX;
 #pragma unroll
     for (int idx = 0; idx < NX; ++idx) {
-      if (idx + thread_offset < num) {
-        in_temp[0] = src[idx];
-        LM2GM(in_temp, dst + idx + thread_offset, sizeof(T));
+      if ((thread_offset + idx) < num) {
+        dst[thread_offset + idx] = src[idx];
       }
     }
-  } else {  // core_num() * NX < num
-    LM2GM(src, dst + thread_offset, NX * sizeof(T));
+  } else {
+    // Vector type
+    constexpr int kVectorSize = (NX % 4 == 0) ? 4 : (NX % 2 == 0) ? 2 : 1;
+    constexpr int kVectorsPerThread = NX / kVectorSize;
+
+    int thread_offset = threadIdx.x * kVectorsPerThread;
+    using VecType = details::VectorType<T, kVectorSize>;
+    VecType* vec_dst = reinterpret_cast<VecType*>(dst);
+    VecType vec_temp[kVectorsPerThread];
+#pragma unroll
+    for (int idx = 0; idx < kVectorsPerThread; ++idx) {
+      vec_temp[idx] = *(reinterpret_cast<VecType*>(src) + idx);
+      vec_dst[thread_offset + idx] = vec_temp[idx];
+    }
   }
 }
 
@@ -403,47 +493,51 @@ __device__ void WriteData(T _global_ptr_* dst, const T* src, int num) {
  *
  * @template paraments
  * Tx: The type of data that needs to be stored in registers.
- * Ty: The type of data stored in the global memory.
+ * Ty: The type of data that stored in the global memory.
  * NX: The number of data columns loaded by each thread.
  * NY: The number of data rows loaded by each thread.
- * BlockSize: Identifies the current device thread index method. For xpu,
- * core_id() is used as the index.
+ * BlockSize: Identifies the current device thread index method. For GPU,
+ * threadIdx.x is used as the thread index. Currently only GPU was supported.
  * IsBoundary: Indicates whether to perform block access storage out-of-bounds
  * judgment. When the number of data processed by the block is less than
- * NX x NY x core_num(), boundary judgment is required to avoid memory access
+ * NX x NY x blockDim.x, boundary judgment is required to avoid memory access
  * crossing the boundary.
  *
  * @param：
- * dst: Data pointer of the current block.
+ * dst: The data pointer of the current block.
  * src: The register pointer of the thread, the size is NX * NY.
- * size_nx: The current block needs to load size_nx columns of data, this
- * parameter will be used when IsBoundary = true.
- * size_ny: The current block needs to load size_ny rows of data. This parameter
- * will be used when IsBoundary = true.
+ * size_nx: The maximum offset of the current block is size_nx elements in the
+ * lowest dimension. The parameters are only calculated when isboundary = true.
+ * size_ny: The maximum offset of the current block is size_ny elements in the
+ * first dimension. The parameters are only calculated when isboundary = true.
  * stride_nx: Each read one element stride stride_nx elements in the last dim.
  * stride_ny: Each read one element stride stride_ny elements in the first dim.
  */
-template <typename Tx, typename Ty, int NX, int NY, int BlockSize,
+template <typename Tx,
+          typename Ty,
+          int NX,
+          int NY,
+          int BlockSize,
           bool IsBoundary = false>
-__device__ __inline__ void WriteData(Ty _global_ptr_* dst, const Tx* src,
-                                     int size_nx, int size_ny, int stride_nx,
-                                     int stride_ny) {
-  int thread_offset = core_id();
+__device__ __forceinline__ void WriteData(Ty* dst,
+                                          const Tx* __restrict__ src,
+                                          int size_nx,
+                                          int size_ny,
+                                          int stride_nx,
+                                          int stride_ny) {
+  int thread_offset = threadIdx.x;
   int left_size_nx = size_nx - thread_offset;
-  __local__ Ty in_temp[1];
 
   // Each branch is added for better performance
-  if (NX == 1 && NY == 1) {
+  if (NX == 1 && NY == 1) {  // for NX == 1 and NY == 1
     if (IsBoundary) {
       if (left_size_nx > 0) {
-        in_temp[0] = static_cast<Ty>(src[0]);
-        LM2GM(in_temp, dst + thread_offset, sizeof(Ty));
+        dst[thread_offset] = static_cast<Ty>(src[0]);
       }
     } else {
-      in_temp[0] = static_cast<Ty>(src[0]);
-      LM2GM(in_temp, dst + thread_offset, sizeof(Ty));
+      dst[thread_offset] = static_cast<Ty>(src[0]);
     }
-  } else if (NX == 1) {
+  } else if (NX == 1) {  // for NX == 1 and NY != 1
 #pragma unroll
     for (int idy = 0; idy < NY; ++idy) {
       if (IsBoundary) {
@@ -451,9 +545,7 @@ __device__ __inline__ void WriteData(Ty _global_ptr_* dst, const Tx* src,
           break;
         }
       }
-
-      in_temp[0] = static_cast<Ty>(src[idy]);
-      LM2GM(in_temp, dst + thread_offset + idy * stride_ny, sizeof(Ty));
+      dst[thread_offset + idy * stride_ny] = static_cast<Ty>(src[idy]);
     }
   } else if (NY == 1) {  // for NY == 1 and NX != 1
 #pragma unroll
@@ -463,9 +555,7 @@ __device__ __inline__ void WriteData(Ty _global_ptr_* dst, const Tx* src,
           break;
         }
       }
-
-      in_temp[0] = static_cast<Ty>(src[idx]);
-      LM2GM(in_temp, dst + thread_offset + idx * stride_nx, sizeof(Ty));
+      dst[thread_offset + idx * stride_nx] = static_cast<Ty>(src[idx]);
     }
   } else {  // for NX != 1 and NY != 1
 #pragma unroll
@@ -482,9 +572,8 @@ __device__ __inline__ void WriteData(Ty _global_ptr_* dst, const Tx* src,
             break;
           }
         }
-        in_temp[0] = static_cast<Ty>(src[idx + idy * NX]);
-        LM2GM(in_temp, dst + thread_offset + idx * stride_nx + idy * stride_ny,
-              sizeof(Ty));
+        dst[thread_offset + idx * stride_nx + idy * stride_ny] =
+            static_cast<Ty>(src[idy * NX + idx]);
       }
     }
   }
@@ -502,7 +591,7 @@ __device__ __inline__ void WriteData(Ty _global_ptr_* dst, const Tx* src,
  * init_data: The register pointer of init data, the size is NX.
  */
 template <typename T, int NX, bool IsBoundary = false>
-__device__ __inline__ void Init(T* dst, T* init_data, int num) {
+__device__ __forceinline__ void Init(T* dst, T* init_data, int num) {
 #pragma unroll
   for (int i = 0; i < NX; i++) {
     if (IsBoundary) {
@@ -521,47 +610,55 @@ __device__ __inline__ void Init(T* dst, T* init_data, int num) {
  * T: The type of data stored in the global memory.
  * NX: The number of data continuously loaded by each thread.
  * NY: The number of data rows loaded by each thread, only NY = 1 was supported.
- * BlockSize: Identifies the current device thread index method. For xpu,
- * core_id() is used as the index.
+ * BlockSize: Identifies the current device thread index method. For GPU,
+ * threadIdx.x is used as the thread index. Currently only GPU was supported.
  * Rank: The shape size of out. eg in[1, 35], out[32, 35] then shape size is 2.
  * IsBoundary: Indicates whether to perform block access storage out-of-bounds
  * judgment. When the number of data processed by the block is less than
- * NX x NY x core_num(), boundary judgment is required to avoid memory access
+ * NX x NY x blockDim.x, boundary judgment is required to avoid memory access
  * crossing the boundary.
  *
  * @param：
  * dst: The register pointer of the thread, the size is NX * NY.
  * src: The original input data pointer of kernel.
- * block_offset: The data offset of this block, core_num() * blockIdx.x * NX;
+ * block_offset: The data offset of this block, blockDim.x * blockIdx.x * NX;
  * config: Calculation configuration of broadcast. It is used to calculate the
  * coordinate mapping relationship between output data and input data.
  * total_num_output: Total number of original output.
  */
-template <typename T, int NX, int NY, int BlockSize, int Rank,
+template <typename T,
+          int NX,
+          int NY,
+          int BlockSize,
+          int Rank,
           bool IsBoundary = false>
-__device__ __inline__ void ReadDataBc(T* dst, const T _global_ptr_* src,
-                                      uint32_t block_offset,
-                                      details::BroadcastConfig<Rank> config,
-                                      int total_num_output) {
-  int thread_offset = block_offset + core_id() * NX;
-  int index_src = 0;
+__device__ __forceinline__ void ReadDataBc(
+    T* dst,
+    const T* __restrict__ src,
+    uint32_t block_offset,
+    details::BroadcastConfig<Rank> config,
+    int total_num_output) {
+  uint32_t thread_offset = block_offset + threadIdx.x * NX;
+  uint32_t index_src = 0;
 
-  __local__ T in_temp;
 #pragma unroll
-  for (int nx = 0; nx < NX; ++nx) {
-    int index_output = thread_offset + nx;
+  for (uint32_t nx = 0; nx < NX; ++nx) {
+    uint32_t index_output = thread_offset + nx;
     index_src = 0;
     if (IsBoundary) {
       if (index_output >= total_num_output) {
         break;
       }
     }
-    index_src = config(index_output);
-    GM2LM(src + index_src, &in_temp, sizeof(T));
-    dst[nx] = in_temp;
+#pragma unroll
+    for (int i = 0; i < Rank; ++i) {
+      auto fast_divmoder = config.divmoders[i].Divmod(index_output);
+      index_output = fast_divmoder.val[0];
+      index_src += fast_divmoder.val[1] * config.strides[i];
+    }
+    dst[nx] = src[index_src];
   }
 }
 
-}  // namespace kernel_primitives
-}  // namespace operators
-}  // namespace paddle
+}  // namespace kps
+}  // namespace pten
